+++ sdk/core/azure-core/azure/core/pipeline/policies/_universal.py
@@ -36,6 +36,7 @@ import types
 import re
 from typing import (Mapping, IO, TypeVar, TYPE_CHECKING, Type, cast, List, Callable, Iterator, # pylint: disable=unused-import
                     Any, Union, Dict, Optional, AnyStr)
+from six.moves import urllib
 
 from azure.core import __version__  as azcore_version
 from azure.core.exceptions import (
@@ -66,7 +67,7 @@ class HeadersPolicy(SansIOHTTPPolicy):
 
     .. admonition:: Example:
 
-        .. literalinclude:: ../examples/test_example_sansio.py
+        .. literalinclude:: ../samples/test_example_sansio.py
             :start-after: [START headers_policy]
             :end-before: [END headers_policy]
             :language: python
@@ -114,7 +115,7 @@ class UserAgentPolicy(SansIOHTTPPolicy):
 
     .. admonition:: Example:
 
-        .. literalinclude:: ../examples/test_example_sansio.py
+        .. literalinclude:: ../samples/test_example_sansio.py
             :start-after: [START user_agent_policy]
             :end-before: [END user_agent_policy]
             :language: python
@@ -130,7 +131,7 @@ class UserAgentPolicy(SansIOHTTPPolicy):
         self.use_env = kwargs.pop('user_agent_use_env', True)
 
         if base_user_agent is None:
-            self._user_agent = "azsdk-python-core/{} Python/{} {}".format(
+            self._user_agent = "azsdk-python-core/{} Python/{} ({})".format(
                 azcore_version,
                 platform.python_version(),
                 platform.platform()
@@ -185,7 +186,7 @@ class NetworkTraceLoggingPolicy(SansIOHTTPPolicy):
 
     .. admonition:: Example:
 
-        .. literalinclude:: ../examples/test_example_sansio.py
+        .. literalinclude:: ../samples/test_example_sansio.py
             :start-after: [START network_trace_logging_policy]
             :end-before: [END network_trace_logging_policy]
             :language: python
@@ -265,6 +266,107 @@ class NetworkTraceLoggingPolicy(SansIOHTTPPolicy):
                         _LOGGER.debug(response.http_response.text())
             except Exception as err:  # pylint: disable=broad-except
                 _LOGGER.debug("Failed to log response: %s", repr(err))
+
+
+class HttpLoggingPolicy(SansIOHTTPPolicy):
+    """The Pipeline policy that handles logging of HTTP requests and responses.
+    """
+
+    DEFAULT_HEADERS_WHITELIST = set([
+        "x-ms-client-request-id",
+        "x-ms-return-client-request-id",
+        "traceparent",
+        "Accept",
+        "Cache-Control",
+        "Connection",
+        "Content-Length",
+        "Content-Type",
+        "Date",
+        "ETag",
+        "Expires",
+        "If-Match",
+        "If-Modified-Since",
+        "If-None-Match",
+        "If-Unmodified-Since",
+        "Last-Modified",
+        "Pragma",
+        "Request-Id",
+        "Retry-After",
+        "Server",
+        "Transfer-Encoding",
+        "User-Agent"
+    ])
+    REDACTED_PLACEHOLDER = "REDACTED"
+
+    def __init__(self, logger=None, **kwargs):  # pylint: disable=unused-argument
+        self.logger = logger or logging.getLogger(
+            "azure.core.pipeline.policies.http_logging_policy"
+        )
+        self.allowed_query_params = set()
+        self.allowed_header_names = set(HttpLoggingPolicy.DEFAULT_HEADERS_WHITELIST)
+
+    def _redact_query_param(self, key, value):
+        lower_case_allowed_query_params = [
+            param.lower() for param in self.allowed_query_params
+        ]
+        return value if key.lower() in lower_case_allowed_query_params else HttpLoggingPolicy.REDACTED_PLACEHOLDER
+
+    def _redact_header(self, key, value):
+        lower_case_allowed_header_names = [
+            header.lower() for header in self.allowed_header_names
+        ]
+        return value if key.lower() in lower_case_allowed_header_names else HttpLoggingPolicy.REDACTED_PLACEHOLDER
+
+    def on_request(self, request):
+        # type: (PipelineRequest) -> None
+        """Logs HTTP method, url and headers.
+        :param request: The PipelineRequest object.
+        :type request: ~azure.core.pipeline.PipelineRequest
+        """
+        http_request = request.http_request
+        options = request.context.options
+        # Get logger in my context first (request has been retried)
+        # then read from kwargs (pop if that's the case)
+        # then use my instance logger
+        logger = request.context.setdefault("logger", options.pop("logger", self.logger))
+
+        if not logger.isEnabledFor(logging.INFO):
+            return
+
+        try:
+            parsed_url = list(urllib.parse.urlparse(http_request.url))
+            parsed_qp = urllib.parse.parse_qsl(parsed_url[4], keep_blank_values=True)
+            filtered_qp = [(key, self._redact_query_param(key, value)) for key, value in parsed_qp]
+            # 4 is query
+            parsed_url[4] = "&".join(["=".join(part) for part in filtered_qp])
+            redacted_url = urllib.parse.urlunparse(parsed_url)
+
+            logger.info("Request URL: %r", redacted_url)
+            logger.info("Request method: %r", http_request.method)
+            logger.info("Request headers:")
+            for header, value in http_request.headers.items():
+                value = self._redact_header(header, value)
+                logger.info("    %r: %r", header, value)
+        except Exception as err:  # pylint: disable=broad-except
+            logger.warning("Failed to log request: %s", repr(err))
+
+    def on_response(self, request, response):
+        # type: (PipelineRequest, PipelineResponse) -> None
+        http_response = response.http_response
+
+        try:
+            logger = response.context["logger"]
+
+            if not logger.isEnabledFor(logging.INFO):
+                return
+
+            logger.info("Response status: %r", http_response.status_code)
+            logger.info("Response headers:")
+            for res_header, value in http_response.headers.items():
+                value = self._redact_header(res_header, value)
+                logger.info("    %r: %r", res_header, value)
+        except Exception as err:  # pylint: disable=broad-except
+            logger.warning("Failed to log response: %s", repr(err))
 
 
 class ContentDecodePolicy(SansIOHTTPPolicy):
@@ -410,7 +512,7 @@ class ProxyPolicy(SansIOHTTPPolicy):
 
     .. admonition:: Example:
 
-        .. literalinclude:: ../examples/test_example_sansio.py
+        .. literalinclude:: ../samples/test_example_sansio.py
             :start-after: [START proxy_policy]
             :end-before: [END proxy_policy]
             :language: python