+++ sdk/identity/azure-identity/azure/identity/_credentials/environment.py
@@ -29,14 +29,15 @@ class EnvironmentCredential:
     a user with a username and password. Configuration is attempted in this order, using these environment variables:
 
     Service principal with secret:
+      - **AZURE_TENANT_ID**: ID of the service principal's tenant. Also called its 'directory' ID.
       - **AZURE_CLIENT_ID**: the service principal's client ID
       - **AZURE_CLIENT_SECRET**: one of the service principal's client secrets
-      - **AZURE_TENANT_ID**: ID of the service principal's tenant. Also called its 'directory' ID.
 
     Service principal with certificate:
-      - **AZURE_CLIENT_ID**: the service principal's client ID
-      - **AZURE_CLIENT_CERTIFICATE_PATH**: path to a PEM-encoded certificate file including the private key
       - **AZURE_TENANT_ID**: ID of the service principal's tenant. Also called its 'directory' ID.
+      - **AZURE_CLIENT_ID**: the service principal's client ID
+      - **AZURE_CLIENT_CERTIFICATE_PATH**: path to a PEM-encoded certificate file including the private key. The
+        certificate must not be password-protected.
 
     User with username and password:
       - **AZURE_CLIENT_ID**: the application's client ID