+++ sdk/keyvault/azure-keyvault-secrets/tests/test_secrets_async.py
@@ -117,10 +117,12 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
         updated = await _update_secret(created)
 
         # delete secret
-        deleted = await client.delete_secret(updated.name)
+        if self.is_playback:
+            polling_interval = 0
+        else:
+            polling_interval = None
+        deleted = await client.delete_secret(updated.name, _polling_interval=polling_interval)
         self.assertIsNotNone(deleted)
-
-        await self._poll_until_exception(client.get_secret, updated.name, expected_exception=ResourceNotFoundError)
 
     @ResourceGroupPreparer(name_prefix=name_prefix)
     @AsyncVaultClientPreparer()
@@ -142,7 +144,7 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
                 expected[secret_name] = secret
 
         # list secrets
-        result = client.list_properties_of_secrets(max_results=max_secrets)
+        result = client.list_properties_of_secrets(max_page_size=max_secrets)
         await self._validate_secret_list(result, expected)
 
     @ResourceGroupPreparer(name_prefix=name_prefix)
@@ -159,14 +161,13 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
             secret_name = "secret{}".format(i)
             secret_value = "value{}".format(i)
             expected[secret_name] = await client.set_secret(secret_name, secret_value)
-
+        if self.is_playback:
+            polling_interval = 0
+        else:
+            polling_interval = None
         # delete them
         for secret_name in expected.keys():
-            await client.delete_secret(secret_name)
-        for secret_name in expected.keys():
-            await self._poll_until_no_exception(
-                client.get_deleted_secret, *expected.keys(), expected_exception=ResourceNotFoundError
-            )
+            await client.delete_secret(secret_name, _polling_interval=polling_interval)
 
         # validate list deleted secrets with attributes
         async for deleted_secret in client.list_deleted_secrets():
@@ -196,7 +197,7 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
                 expected[secret.id] = secret
 
         # list secret versions
-        result = client.list_secret_versions(secret_name)
+        result = client.list_properties_of_secret_versions(secret_name)
 
         # validate list secret versions with attributes
         async for secret in result:
@@ -223,7 +224,11 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
         self.assertIsNotNone(secret_backup, "secret_backup")
 
         # delete secret
-        await client.delete_secret(created_bundle.name)
+        if self.is_playback:
+            polling_interval = 0
+        else:
+            polling_interval = None
+        await client.delete_secret(created_bundle.name, _polling_interval=polling_interval)
 
         # restore secret
         restored = await client.restore_secret_backup(secret_backup)
@@ -246,11 +251,12 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
             secrets[secret_name] = await client.set_secret(secret_name, secret_value)
 
         # delete all secrets
+        if self.is_playback:
+            polling_interval = 0
+        else:
+            polling_interval = None
         for secret_name in secrets.keys():
-            await client.delete_secret(secret_name)
-        await self._poll_until_no_exception(
-            client.get_deleted_secret, *secrets.keys(), expected_exception=ResourceNotFoundError
-        )
+            await client.delete_secret(secret_name, _polling_interval=polling_interval)
 
         # validate all our deleted secrets are returned by list_deleted_secrets
         async for deleted_secret in client.list_deleted_secrets():
@@ -258,7 +264,7 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
 
         # recover select secrets
         for secret_name in secrets.keys():
-            await client.recover_deleted_secret(secret_name)
+            await client.recover_deleted_secret(secret_name, _polling_interval=polling_interval)
 
         # validate the recovered secrets exist
         await self._poll_until_no_exception(
@@ -281,11 +287,12 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
             secrets[secret_name] = await client.set_secret(secret_name, secret_value)
 
         # delete all secrets
+        if self.is_playback:
+            polling_interval = 0
+        else:
+            polling_interval = None
         for secret_name in secrets.keys():
-            await client.delete_secret(secret_name)
-        await self._poll_until_no_exception(
-            client.get_deleted_secret, *secrets.keys(), expected_exception=ResourceNotFoundError
-        )
+            await client.delete_secret(secret_name, _polling_interval=polling_interval)
 
         # validate all our deleted secrets are returned by list_deleted_secrets
         async for deleted_secret in client.list_deleted_secrets():